SonarCloud problems gefixed
