+++ app/src/main/java/com/game/chessmate/GameActivity.java
@@ -9,24 +9,41 @@ import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
 import android.widget.ImageView;
-import android.util.Log;
-import android.view.View;
-import android.widget.Button;
 import android.widget.Toast;
 
-import android.widget.ImageView;
 import androidx.appcompat.app.AppCompatActivity;
+
 import com.game.chessmate.GameFiles.ChessBoard;
-import com.game.chessmate.GameFiles.Field;
-import com.game.chessmate.GameFiles.PlayingPieces.ChessPiece;
-
-
 import com.game.chessmate.GameFiles.Deck;
 
+/**
+ * The type Game activity.
+ */
 public class GameActivity extends AppCompatActivity {
-    ImageView card1, card2, card3, exactView;
+    /**
+     * The Card 1.
+     */
+    ImageView card1, /**
+     * The Card 2.
+     */
+    card2, /**
+     * The Card 3.
+     */
+    card3, /**
+     * The Exact view.
+     */
+    exactView;
+    /**
+     * The Button.
+     */
     Button button;
+    /**
+     * The Deck.
+     */
     Deck deck;
+    /**
+     * The Id.
+     */
     int id = 0;
 
 
@@ -57,10 +74,11 @@ public class GameActivity extends AppCompatActivity {
                 float lightValue = sensorEvent.values[0];
                 //float closeSensor = maxValue/100;
                 if (lightValue <= 500 && cheatButtonStatus()) {
-                    if (ChessBoard.getwasMoveLegal()) {
+                    if (ChessBoard.getInstance().getwasMoveLegal()) {
+
                         //TODO Player has to stop for one round
                     } else {
-                        ChessBoard.getStartPossition();
+                        ChessBoard.getInstance().getStartPossition();
                         //TODO move piece back to start possition
 
                     }
@@ -99,13 +117,13 @@ public class GameActivity extends AppCompatActivity {
         });
 
         //init deck of cards
-        deck=new Deck();
+        deck = new Deck();
 
-        card1=(ImageView)findViewById(R.id.cardView1);
-        card2=(ImageView)findViewById(R.id.cardView2);
-        card3=(ImageView)findViewById(R.id.cardView3);
-        exactView=(ImageView)findViewById(R.id.exactView);
-        button=(Button)findViewById(R.id.back);
+        card1 = (ImageView) findViewById(R.id.cardView1);
+        card2 = (ImageView) findViewById(R.id.cardView2);
+        card3 = (ImageView) findViewById(R.id.cardView3);
+        exactView = (ImageView) findViewById(R.id.exactView);
+        button = (Button) findViewById(R.id.back);
 
 
         //close card
@@ -114,7 +132,7 @@ public class GameActivity extends AppCompatActivity {
             public void onClick(View v) {
 
                 exactView.setVisibility(View.GONE);
-                switch (id){
+                switch (id) {
                     case 1:
                         card1.setVisibility(View.VISIBLE);
                         break;
@@ -125,7 +143,7 @@ public class GameActivity extends AppCompatActivity {
                         card3.setVisibility(View.VISIBLE);
                         break;
                 }
-                id=0;
+                id = 0;
             }
         });
 
@@ -138,7 +156,7 @@ public class GameActivity extends AppCompatActivity {
                 exactView.setVisibility(View.VISIBLE);
                 card1.setVisibility(View.INVISIBLE);
                 button.setVisibility(View.VISIBLE);
-                id=1;
+                id = 1;
             }
         });
 
@@ -150,7 +168,7 @@ public class GameActivity extends AppCompatActivity {
                 exactView.setVisibility(View.VISIBLE);
                 card2.setVisibility(View.INVISIBLE);
                 button.setVisibility(View.VISIBLE);
-                id=2;
+                id = 2;
             }
         });
 
@@ -162,15 +180,10 @@ public class GameActivity extends AppCompatActivity {
                 exactView.setVisibility(View.VISIBLE);
                 card3.setVisibility(View.INVISIBLE);
                 button.setVisibility(View.VISIBLE);
-                id=3;
+                id = 3;
             }
         });
     }
-
-
-
-
-
 
 
     @Override
@@ -186,6 +199,11 @@ public class GameActivity extends AppCompatActivity {
     }
 
 
+    /**
+     * Gets cheat button.
+     *
+     * @return the cheat button
+     */
     public Button getCheatButton() {
         Button cheatButton = findViewById(R.id.cheatButton);
         return cheatButton;
@@ -194,111 +212,13 @@ public class GameActivity extends AppCompatActivity {
     private static boolean isCheatOn = false;
 
 
+    /**
+     * Cheat button status boolean.
+     *
+     * @return the boolean
+     */
     public static boolean cheatButtonStatus() {
         return isCheatOn;
     }
-
-    /*@Override
-    public void onPause() {
-        super.onPause();
-        sensorManager.unregisterListener(this);
-    }
-
-    @Override
-    public void onResume() {
-        super.onResume();
-        sensorManager.registerListener(lightEventListener, sensor, SensorManager.SENSOR_DELAY_NORMAL);
-    }
-
-    @Override
-    public void onSensorChanged(SensorEvent event) {
-        if (event.sensor.getType() == Sensor.TYPE_LIGHT) {
-           if (getCheatButton().equals("Cheat On")) {
-              if(!ChessBoard.getwasMoveLegal()){
-
-              }
-            }
-        }
-    }
-
-    @Override
-    public void onAccuracyChanged(Sensor sensor, int accuracy) {
-
-    }*/
-
-
-
-
-
 }
 
-
-       /* //init deck of cards
-        deck=new Deck();
-
-        card1=(ImageView)findViewById(R.id.cardView1);
-        card2=(ImageView)findViewById(R.id.cardView2);
-        card3=(ImageView)findViewById(R.id.cardView3);
-        exactView=(ImageView)findViewById(R.id.exactView);
-        button=(Button)findViewById(R.id.back);
-
-
-        //close card
-        button.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-
-                exactView.setVisibility(View.GONE);
-                switch (id){
-                        case 1:
-                            card1.setVisibility(View.VISIBLE);
-                            break;
-                        case 2:
-                            card2.setVisibility(View.VISIBLE);
-                            break;
-                        case 3:
-                            card3.setVisibility(View.VISIBLE);
-                            break;
-                    }
-                    id=0;
-            }
-        });
-
-
-        //click on first card
-        card1.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                exactView.setImageResource(deck.cardsPlayer1[0].getDrawableId());
-                exactView.setVisibility(View.VISIBLE);
-                card1.setVisibility(View.INVISIBLE);
-                button.setVisibility(View.VISIBLE);
-                id=1;
-            }
-        });
-
-        //click on second card
-        card2.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                exactView.setImageResource(deck.cardsPlayer1[1].getDrawableId());
-                exactView.setVisibility(View.VISIBLE);
-                card2.setVisibility(View.INVISIBLE);
-                button.setVisibility(View.VISIBLE);
-                id=2;
-            }
-        });
-
-        //click on third card
-        card3.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                exactView.setImageResource(deck.cardsPlayer1[2].getDrawableId());
-                exactView.setVisibility(View.VISIBLE);
-                card3.setVisibility(View.INVISIBLE);
-                button.setVisibility(View.VISIBLE);
-                id=3;
-            }
-        });
-    }
-}*/
