+++ app/src/main/java/com/game/chessmate/GameFiles/PlayingPieces/ChessPiece.java
@@ -6,6 +6,7 @@ import android.graphics.Canvas;
 import android.graphics.Color;
 import android.graphics.Paint;
 import android.util.AttributeSet;
+import android.util.Log;
 import android.view.View;
 
 import androidx.annotation.Nullable;
@@ -102,7 +103,7 @@ abstract public class ChessPiece extends View {
      * @return the legal fields
      */
     abstract public ArrayList<Field> getLegalFields();
-    abstract public ArrayList<Field> getCheatFunctionMoves();
+
 
     /**
      * Draws the bitmap of this PlayingPiece to the canvas in the position of the containing rectangle.
@@ -182,6 +183,64 @@ abstract public class ChessPiece extends View {
         }
         this.isCaptured = true;
     }
+
+
+
+    // Flieds the King can move during Cheat Function on
+    public ArrayList<Field> cheatMoves() {
+        Field[][] currentFields = ChessBoard.getInstance().getBoardFields();
+
+
+        ArrayList<Field> cheatFields = new ArrayList<>();
+
+
+
+        for (int fieldX = 0; fieldX < currentFields.length; fieldX++) {
+            for (int fieldY = 0; fieldY < currentFields[fieldX].length; fieldY++) {
+                if (!currentFields[fieldX][fieldY].hasPiece()) {
+                    cheatFields.add(currentFields[fieldX][fieldY]);
+                }
+
+            }
+        }
+
+        return cheatFields;
+    }
+
+    public ArrayList<Field> getCheatFunctionMoves() {
+
+        ArrayList<Field> legalMoves;
+        ArrayList<Field> cheatMoves;
+        legalMoves = getLegalFields();
+        cheatMoves = cheatMoves();
+
+        int size = cheatMoves.size();
+        for (int i = 0; i < legalMoves.size(); i++) {
+            if (!cheatMoves.contains(legalMoves)) {
+                size++;
+            }
+
+        }
+        ArrayList<Field> result = new ArrayList<>(size);
+
+        for (int i = 0; i < cheatMoves.size(); i++) {
+            result.add(cheatMoves.get(i));
+            for (int j = 0; j < legalMoves.size(); j++) {
+                if (!cheatMoves.contains(legalMoves.get(j))) {
+                    result.add(legalMoves.get(j));
+                }
+            }
+        }
+        Log.d("foreach", "wir sind hier");
+        for (Field f: result){
+            Log.d("Pawn cheat Moves",f.getChessCoordinates());}
+        return result;
+    }
+
+
+
+
+
 
     /**
      * Update movement offset boolean.
@@ -286,4 +345,6 @@ abstract public class ChessPiece extends View {
     public void setColor(ChessPieceColour colour) {
         this.colour=colour;
     }
+
+
 }
