+++ app/src/main/java/com/game/chessmate/GameFiles/PlayingPieces/Pawn.java
@@ -3,7 +3,6 @@ package com.game.chessmate.GameFiles.PlayingPieces;
 import android.content.Context;
 import android.graphics.Bitmap;
 import android.util.AttributeSet;
-import android.util.Log;
 
 import androidx.annotation.Nullable;
 
@@ -56,58 +55,5 @@ public class Pawn extends ChessPiece {
     }
 
 
-    public ArrayList<Field> cheatMoves() {
-        Field[][] currentFields = ChessBoard.getInstance().getBoardFields();
-       /* for (Field[] f: currentFields){
-            Log.d("currentFields", f.toString());
-        }
-        String size = String.valueOf(currentFields.length);
-        Log.d("CURRENTFIELDSIZE", size); // 8*/
 
-        ArrayList<Field> cheatFields = new ArrayList<>();
-       // int i = currentPosition.getFieldX();
-        //int j = currentPosition.getFieldY();
-
-
-        for (int fieldX = 0; fieldX < currentFields.length; fieldX++) {
-            for (int fieldY = 0; fieldY < currentFields[fieldX].length; fieldY++) {
-                if (!currentFields[fieldX][fieldY].hasPiece()) {
-                    cheatFields.add(currentFields[fieldX][fieldY]);
-                }
-
-            }
-        }
-       // Log.d("CHEATFIELD>SIZE", cheatFields.toString()) ;
-        return cheatFields;
-    }
-
-    public ArrayList<Field> getCheatFunctionMoves() {
-
-        ArrayList<Field> legalMoves = new ArrayList<>();
-        ArrayList<Field> cheatMoves = new ArrayList<>();
-        legalMoves = getLegalFields();
-        cheatMoves = cheatMoves();
-
-        int size = cheatMoves.size();
-        for (int i = 0; i < legalMoves.size(); i++) {
-            if (!cheatMoves.contains(legalMoves)) {
-                size++;
-            }
-
-        }
-        ArrayList<Field> result = new ArrayList<>(size);
-
-        for (int i = 0; i < cheatMoves.size(); i++) {
-            result.add(cheatMoves.get(i));
-            for (int j = 0; j < legalMoves.size(); j++) {
-                if (!cheatMoves.contains(legalMoves.get(j))) {
-                    result.add(legalMoves.get(j));
-                }
-            }
-        }
-        Log.d("foreach", "wir sind hier");
-        for (Field f: result){
-        Log.d("Pawn cheat Moves",f.getChessCoordinates());}
-        return result;
-    }
 }
